:recycle: [fix] #87 뱃지 추가 및 패키지 이동
