+++ src/main/java/com/sopt/bbangzip/domain/badge/api/dto/response/BadgeResponse.java
@@ -1,7 +1,10 @@
 package com.sopt.bbangzip.domain.badge.api.dto.response;
 
+import lombok.Builder;
+
 import java.util.List;
 
+@Builder
 public record BadgeResponse(
         String badgeName,
         String badgeImage,
